Fix comparison of identical data trees
Use null pointers to represent empty data trees, which is what
the libyang2 API expects from users. DataTree::new() was using
lyd_validate_all() to create semi-empty data trees, but that was
problematic to represent empty data diffs, which are different in a
few ways compared to regular data trees. As an example, the lack of de
facto empty data trees was leading to crashes when comparing identical
data trees (because of missing metadata in the implicit nodes created
by lyd_validate_all()). This commit fixes this problem.

While here, change DataTree::new_path() to return the last
created/modified node instead of the first one (which is generally
less useful).

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>