Use reference counted type to simplify lifetime management
Change the DataTree structure to contain a reference-counting pointer
instead of a regular reference to the associated YANG context. With
this change, the DataTree structure is no longer bound to any
lifetime, improving the API ergonomics and adding more flexibility
(e.g. possibility to send data trees through channels). The only
downside is that YANG contexts can't be modified once wrapped inside
an Arc, but that shouldn't be a limitation for most use cases. This
is a breaking API change that might be revisited in the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>