Fix bug in DataNodeRef::value_canonical()
The libyang documentation says that the lyd_value's _canonical field
should never be accessed directly. Instead, the lyd_get_value()
function should be used for safe access to the canonical
value. Unfortunately lyd_get_value() is an inline function, so we
can't use it in yang2-rs. We can however replicate what it does,
which fixes the problem of DataNodeRef::value_canonical() returning
None when it shouldn't for certain data types.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>