Rework the bundled feature flag
yang-rs with bundled feature is building a shared libyang object, not
a static one as expected by this feature. This commit reworks the
implementation of bundled flag to use only cmake for the build
with the correct flags to generate a static object `.a`. Also,
remove the manual build steps using `cc`.
The bindgen feature now will regenerate bindings based on the
newly compiled code if the bundled feature is enabled.