Allow schema modules to be parsed directly from either a string or a file
Provide a simpler interface to only parse YANG schemas without having to use
Context::load_module, since the `load_module` will fail if there is dependencies
that are not loaded.

This good for use cases where user is interested on checking schema syntax correctness,
or iterating over the schema elements for information.